requirements.txt
yt-dlp
SpeechRecognition
pydub
google-genai
pocketsphinx
fastapi
uvicorn
python-dotenv